Implemented page update without refresh
Tested updating timestamps at a certain time interval. There is
currently a bug when pressing the getJSON button. It continues to
increase the frequency of the setTimout() function. I will continue
further investigation and attempt to make the time stamp incrementation
more modular. I also need to increase the days, months, and years.