Reading JSON keys dynamically and various backend upgrades
- Read in timestamps dynamically and snap time stamps that aren't
exactly at 10 minute increments. It works by comparing the difference
between the actual timestamp sent from the pie and the exact 10 minute
increment. If the difference is less than 10 minutes, it will use that
time stamp that was sent the pi as the timestamp to display.

- Need to account for is the case where data in the json file for a
specific data reading is 0.

- Code clean up and organization

- Still need to dynamically assign the initial timestamp based on the
most recent data in the json file